+++ TCZApp/Views/Components/ImageCropView.swift
@@ -1,0 +1,46 @@
+import SwiftUI
+import CropViewController
+
+/// A SwiftUI wrapper for TOCropViewController with circular cropping for profile pictures.
+struct ImageCropView: UIViewControllerRepresentable {
+    let image: UIImage
+    let onCrop: (UIImage) -> Void
+    let onCancel: () -> Void
+
+    func makeUIViewController(context: Context) -> CropViewController {
+        let cropViewController = CropViewController(croppingStyle: .circular, image: image)
+        cropViewController.delegate = context.coordinator
+        cropViewController.aspectRatioPreset = .presetSquare
+        cropViewController.aspectRatioLockEnabled = true
+        cropViewController.resetAspectRatioEnabled = false
+        cropViewController.rotateButtonsHidden = true
+        cropViewController.rotateClockwiseButtonHidden = true
+        cropViewController.doneButtonTitle = "Fertig"
+        cropViewController.cancelButtonTitle = "Abbrechen"
+        return cropViewController
+    }
+
+    func updateUIViewController(_ uiViewController: CropViewController, context: Context) {}
+
+    func makeCoordinator() -> Coordinator {
+        Coordinator(self)
+    }
+
+    class Coordinator: NSObject, CropViewControllerDelegate {
+        let parent: ImageCropView
+
+        init(_ parent: ImageCropView) {
+            self.parent = parent
+        }
+
+        func cropViewController(_ cropViewController: CropViewController, didCropToCircularImage image: UIImage, withRect cropRect: CGRect, angle: Int) {
+            parent.onCrop(image)
+        }
+
+        func cropViewController(_ cropViewController: CropViewController, didFinishCancelled cancelled: Bool) {
+            if cancelled {
+                parent.onCancel()
+            }
+        }
+    }
+}